+++ src/miniRT/is_white.c
@@ -35,10 +35,10 @@ void	draw_white(t_win w)
 	int			wx0;
 	int			wy0;
 
-	wy0 = w.lu->e[1];
+	wy0 = (int)w.lu->e[1];
 	while (wy0 < w.rd->e[1])
 	{
-		wx0 = w.lu->e[0];
+		wx0 = (int)w.lu->e[0];
 		while (wx0 < w.rd->e[0])
 		{
 			win_pixel_put(w, wx0, wy0, WHITE);
